// src/app/services/models/messageDto.ts
export interface MessageDto {
    messageId: string;
    message: string;
    createdOn: string;
    deleteAfter: string;
    to: string;
    subject: string;
    from: string;
    userId: string;
    userName: string;
    userLoginName: string;
    toUserId: string;
    fromUserId: string;
    emailId: string;
    groupId: string;
    sCode: string;
    rid: string;
  }